♻ Refactor: 팀 페이지 타 입정리/분리
