+++ src/app/(service)/[teamid]/types/index.ts
@@ -1,0 +1,4 @@
+export * from '@/app/(service)/[teamid]/types/common';
+export * from '@/app/(service)/[teamid]/types/member';
+export * from '@/app/(service)/[teamid]/types/task';
+export * from '@/app/(service)/[teamid]/types/team';
